Major updates to Flume and Drill Views
Flume updated version to support a filter for languages and also file
management to allow live feed and query of data with no errors on inUse
Flume files.
Drill views updated to handle suppression of character sets better and
work better with larger volume of data and different characters.
Documentation need to be updated still.